test: move globalShortcut spec
